Added changes that are live (#1)

+++ application/views/master/view_order_details.php
@@ -3,8 +3,11 @@
             <div class="container-fluid" id="print_view">
                     <p>
                         <a class="btn btn-primary" href="<?= site_url(); ?>/order_print_view/<?= $this->uri->segment(2, 0); ?>">Print</a>
+                        
+                        <?php if (isset($result[0]['ostatus']) && $result[0]['ostatus'] == 'In Process') { ?>
                         <a   href="<?= site_url(); ?>/edit_order/<?= $this->uri->segment(2, 0); ?>" class="btn btn-info" >Add Product</a>
                         <a class="btn btn-danger" href="<?= site_url(); ?>/cancel_order/<?= $this->uri->segment(2, 0); ?>"   onclick="return onDeleteItem()">Cancel Order</a>
+                        <?php } ?>
                     </p>
             </div>
             <div class="container-fluid" id="print_view">
@@ -48,7 +51,10 @@
                             </tr>
                             <tr>
                                 <td><strong>Customer Address</strong></td>
-                                <td><?php if (isset($result[0]['customer_address'])) echo $result[0]['customer_address'] ?></td>    
+                                <td><?php if (isset($result[0]['house_no'])) 
+                                                echo $result[0]['house_no']." ".$result[0]['street_address'].", ".$result[0]['apartment_name'].", ".$result[0]['landmark'].", ".$result[0]['pincode'].", ".$result[0]['city']
+                                    ?>                
+                                </td> 
                             </tr>
                                         
                             
@@ -86,7 +92,7 @@
                                                     <td> <?=  $value['discount_amount']; ?></td>       
                                                     <td> <?=  $value['item_total_amount']; ?></td> 
                                                     <td> 
-                                                        <?php if($value['active'] == '1') { ?>
+                                                        <?php if($value['active'] == '1' &&  $result[0]['ostatus'] == 'In Progress') { ?>
                                                             <a href="<?= site_url(); ?>/delete_order_item/<?= $value['order_item_id']; ?>"  class="btn btn-danger" name="remove<?= $count?>" id="remove<?= $count?>" onclick="return onDeleteItem()" >Delete</a>
                                                         <?php } ?>
                                                     </td>
